Setup: wired configuration overwrite functionality in app
